Claim support of gchat:// scheme in the desktop file and process them
When generating XDG Desktop file claim support of gchat:// scheme.
If an URL with this scheme is passed on the command line replace that
scheme with https:// one and navigate to that URL. This works for both
new instance and when an instance is already running - for
"second-instance".

Once this is installed in the system and one navigates to URL with this
scheme in a browser it offers open of this URL with this application.

Last piece is to rewrite Google Chat links (that use https:// scheme)
with the psedo-protocol gchat:// that allows all of the above.